pyproject.toml
[tool.poetry]
name = "twooter"
version = "0.1.0"
description = "Interview assignment for Northern Tech"
authors = ["Tomas Sandven <tomas@sandven.email>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.97.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.3"}
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
